requirements.txt
streamlit
openai
langchain
pytesseract
Pillow
pandas
python-dotenv
fastapi
uvicorn[standard]
flask
opencv-python
werkzeug
easyocr
torch
firebase-admin
python-dotenv
fastapi
uvicorn
requests

